refactorisation

+++ modif_article.php
@@ -23,10 +23,6 @@ if(!isset($_SESSION["id"]) || $_SESSION["id"] != $article['id_users']){
 <head>
         <meta charset="UTF-8">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
-
-          <!-- lien bootstrap  -->
-        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-KK94CHFLLe+nY2dmCWGMq91rCGa5gtU4mk92HdvYe+M/SXH301p5ILy+dN9+nJOZ" crossorigin="anonymous">
-  
         <title>Modifier un article</title>
 </head>
 <body>
@@ -39,95 +35,30 @@ if(!isset($_SESSION["id"]) || $_SESSION["id"] != $article['id_users']){
     if(!empty($_SESSION['id'])){
 ?>
 
+<!-- Modal de confirmation de suppression d'article, script modal.js -->
+<div id="dialog-confirm" class="modal fade" tabindex="-1" role="dialog" aria-labelledby="dialog-confirm-title" aria-hidden="true">
+  <div class="modal-dialog">
+    <div class="modal-content">
+      <div class="modal-header">
+        <h5 class="modal-title" id="dialog-confirm-title">Confirmation de la suppression</h5>
+        <button type="button" class="close" data-bs-dismiss="modal" aria-label="Close">
+          <span aria-hidden="true">&times;</span>
+        </button>
+      </div>
+      <div class="modal-body">
+        <p>Etes-vous sûr de vouloir supprimer cet élément ?</p>
+      </div>
+      <div class="modal-footer">
+        <button type="button" class="btn btn-primary" id="confirm-yes">Confirmer</button>
+        <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Annuler</button>
+      </div>
+    </div>
+  </div>
+</div>
 
- <?php 
-
-        if(isset($_GET['reg_err']))
-        {
-            $err = htmlspecialchars($_GET['reg_err']);
-
-            switch($err)
-            {
-                case 'success':
-                ?>
-                    <div class="alert alert-success">
-                        <strong>Succès</strong> Article posté !
-                    </div>
-                <?php
-                break;
-
-                case 'error':
-                ?>
-                    <div class="alert alert-danger">
-                        <strong>Erreur</strong> Erreur
-                    </div>
-                <?php
-                break;
-
-                case 'type_file':
-                ?>
-                    <div class="alert alert-danger">
-                        <strong>Erreur</strong> Image trop lourde ou format incorrect 
-                    </div>
-                <?php
-                break;
-
-                case 'image':
-                ?>
-                    <div class="alert alert-danger">
-                        <strong>Erreur</strong> Merci de sélectionner une image
-                    </div>
-                <?php 
-                break;
-
-                case 'type':
-                ?>
-                    <div class="alert alert-danger">
-                        <strong>Erreur</strong> Type non autorisé. Veuillez choisir une image.png ou .jpg ou .jpeg ou .gif
-                    </div>
-                <?php 
-                break;
-
-                case 'check':
-                ?>
-                    <div class="alert alert-danger">
-                        <strong>Erreur</strong> Veuillez sélectionner une image
-                    </div>
-                <?php 
-                break;
-
-                case 'tit_empty':
-                    ?>
-                        <div class="alert alert-danger">
-                            <strong>Erreur</strong> Veuillez remplir le titre de l'article
-                        </div>
-                    <?php 
-                break;
-
-                case 'cont_empty':
-                    ?>
-                        <div class="alert alert-danger">
-                            <strong>Erreur</strong> Veuillez mettre une description, ou un commentaire à votre article
-                        </div>
-                    <?php 
-                break;
-
-                case 'error_id':
-                    ?>
-                        <div class="alert alert-danger">
-                            <strong>Erreur</strong> Erreur d'id
-                        </div>
-                    <?php 
-                break;
-
-
-            }
-        }
-?>
 
 <?php
    
-
     
     echo     ' <div class="container mt-5 mb-5 col-sm12 col-md-8 col-lg-10 border border-info shadow-lg bg-info-rounded">
                     <div class="row text-center border border-info shadow-lg bg-info-rounded">
@@ -156,7 +87,7 @@ if(!isset($_SESSION["id"]) || $_SESSION["id"] != $article['id_users']){
 
                                         <div class="row">
                                             <div class="d-flex justify-content-around mx-auto mt-5 mb-4">
-                                                <a href="functions/delete_article.php?id= ' . $_GET['id'] . '"<button class="btn btn-info" type="submit">Supprimer entièrement l\'article</button></a>
+                                                <a href="functions/delete_article.php?id= ' . $_GET['id'] . '"<button class="confirmModal btn btn-info" type="submit">Supprimer entièrement l\'article</button></a>
                                                 <button class="btn btn-info" type="submit">Envoyer les modifications</button>
                                             </div>
                                         </div>
@@ -189,9 +120,6 @@ if(!isset($_SESSION["id"]) || $_SESSION["id"] != $article['id_users']){
  <!-- lien footer -->
 <?php require_once 'footer.php' ?>
 
-<!-- lien bootstrap -->
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ENjdO4Dr2bkBIFxQpeoTz1HIcje39Wm4jDKdf19U8gI4ddQ3GYNS7NTKfAdVQSZe" crossorigin="anonymous"></script>
- 
  </body>
 </html>
